Write test for format number
Changed format_number to use .17g. The 17 decimal digits are the required number of significant decimal digits to adequately represent a 64-bit floating point number according to IEEE754-1985. The g specifier uses 17 significant digits rather than 17 decimal digits like the f specifier does.

The .17g specifier is what the Teem NRRD implementation uses. Best to stay consistent.

Note that many times the output float number will be something like 0.10000000000000001 rather than the expected 0.1 but that is due to an error with floating point representations and not the implementation itself.