Compare test Numpy arrays using Numpy.testing module
The previous method was converting to a byte string and comparing the strings. The main issue is that if the assertion fails, there will be no useful information in the trace about the data because it's one large string. By using Numpy's testing module, it will report statistics about how far off the matrix is, size of matrix and where the values are mismatched.