Update setup.py to new file structure
Additionally, setuptools is REQUIRED to package this module now because find_packages is used. Previously, setuptools would attempt to be imported and if not available, this module would fallback to distutils. This is not the case anymore. Unless there is a situation where distutils is necessary, this should be fine.

Instead of having a standalone module that contains one file, there are now multiple files in a folder that must be included.