Fix issue with test on format_numbers failing
Issue was with using repr on Python 3.1 onwards which rounds floating point numbers to 12 decimals. The format_number function rounds to 17 digits which is what is required for full precision of a 64-bit floating point number.

See more information here: https://docs.python.org/3.4/tutorial/floatingpoint.html

The test was altered to use the decimal library and rounding to 17 points instead of using repr.