Capitalize all instances of NRRD
As part of reformatting, all instances of NRRD are capitalized because it is an acronym standing for Nearly Raw Raster Data (NRRD). The exception if referring to the module name, which is lowercase pynrrd and nrrd. Also, if referring to the file extension, it is lowercased to nrrd as well.