Update README.rst with one from documentation and symlink each other
Symlink the README.rst in the documentation to the main README.rst of the repository. I originally thought I had done this but when I edited the documentation README.rst, it did not transfer over to the repository README.rst. Copy changes from the documentation README into the main README.