Create tests that fail with data that is not writeable after being read
Not all data can be written when imported from a Numpy array. In this instance, it seems that any compressed data is unable to be written after it is read.

The Numpy array has a flags variable containing information about the data. One flag, denoted WRITEABLE, indicates whether the data can be written.