Fix key-value writing: no space after ':=' identifier (#69)
per the NRRD documentation: http://teem.sourceforge.net/nrrd/format.html#general.2

"""
Each of the "<key>:=<value>" lines specifies a key/value pair in the nrrd. These can appear in NRRD0002 (and higher version) files, but not NRRD0001 files. The key and value strings are delimited by the first ":=" to appear on the line: any spaces before or after ":=" are assumed part of the key or value, respectively.
"""