Small grammar & wording changes
Great work! For documentation, the only way for me to 'test' that it is okay is by retrieving the PR and compiling it by hand. Since I'm that far, I decided to make some changes of my own and commit them. For the most part, small wording and grammar changes were made.

Also updated the API reference to show the ALLOW_DUPLICATE_FIELD docstring in addition to it being displayed in the table of contents for the reading section.

Update references to True/False & ALLOW_DUPLICATE_FIELD so Sphinx appropriately links to these items.

Also found a small mistake in the tests that was missed originally I decided to fix here. Wasn't worth the extra effort of a separate PR.