Remove documentation regarding compression_level=0 (#91)
The GZIP library supports a compression level of `0` for no compression but this has been found to be slower than setting the encoding to `raw`. This PR removes documentation for setting `compression_level=0`. In addition, `bz2` in Python does not support a compression level of 0. Another benefit of this PR is the unified compression level range.

Note that the user can still specify a compression level of 0 at their own risk.

Fixes issue #88