Add quotes around string lists for labels, units and space units (#98)
The NRRD spec states that the headers "labels", "units" and "space units" should have each item in the string list be delineated with quotes. This is to allow spaces in each of the string items without causing problems.

This PR fixes and adds quotes around the respective header items. In 3D Slicer, if the items were not surrounded with quotes, the NRRD would fail to load.