Change logic for how headers and detached data are written
Ignore 'data file' and 'datafile' fields in the header.

Allow user to specify a custom data path via the detached_header
parameter. If its a string, then it's assumed to be a custom path.

Otherwise, if detached_header is set to True, then the filename for the
data will be calculated.