Fix tests failing in CI/CD (#166)
Pin Ubuntu version to 22.04 for all workflows. This fixes the issue with tests failing on Python 3.7 due to Python 3.7 not being available on the latest Ubuntu OS (see here: https://github.com/actions/setup-python/issues/962). Pinning the OS version on **all** workflows should limit unforeseen errors due to an a new latest OS.

Update Python version used in lint/publish workflows to 13. This is the latest available for Ubuntu 22.04.

Add Python 3.12 version to tests for CI. **Note:** Tests fail when testing with Python 3.13. This will be fixed in an upcoming PR.

Add Python 3.12/3.13 to list of supported versions.