Stop closing files opened by the caller (#165)
Fixes mhe/pynrrd#164

Note: this was a very quick edit on a Friday afternoon before I log off for the week.

I have only run through a test in the case that the file is passed in (fh_is_managed = True), not in the other case (fh_is_managed = False).

Co-authored-by: Jack Riccomini <j.riccomini@qureight.com>
Co-authored-by: Addison Elliott <addison.elliott@gmail.com>